// PTAngular/projectupdatesfrontend/src/app/resources/resource-utility.ts
export interface IResourceUtilization {

  resourceUtilizationSummaryId: string;
  year: number;
  companyId: string;
  resourceId: string;
  janResourceUtilizationInDays: number;
  febResourceUtilizationInDays: number;
  marResourceUtilizationInDays: number;
  aprResourceUtilizationInDays: number;
  mayResourceUtilizationInDays: number;
  junResourceUtilizationInDays: number;
  julResourceUtilizationInDays: number;
  augResourceUtilizationInDays: number;
  sepResourceUtilizationInDays: number;
  octResourceUtilizationInDays: number;
  novResourceUtilizationInDays: number;
  decResourceUtilizationInDays: number;
  totalUtilizationInDays: number;

  janAvailabilityBeforeHolidaysInDays: number;
  febAvailabilityBeforeHolidaysInDays: number;
  marAvailabilityBeforeHolidaysInDays: number;
  aprAvailabilityBeforeHolidaysInDays: number;
  mayAvailabilityBeforeHolidaysInDays: number;
  junAvailabilityBeforeHolidaysInDays: number;
  julAvailabilityBeforeHolidaysInDays: number;
  augAvailabilityBeforeHolidaysInDays: number;
  sepAvailabilityBeforeHolidaysInDays: number;
  octAvailabilityBeforeHolidaysInDays: number;
  novAvailabilityBeforeHolidaysInDays: number;
  decAvailabilityBeforeHolidaysInDays: number;

  totalAvailabilityBeforeHolidaysInDays: number;
  janTotalHolidays: number;
  febTotalHolidays: number;
  marTotalHolidays: number;
  aprTotalHolidays: number;
  mayTotalHolidays: number;
  junTotalHolidays: number;
  julTotalHolidays: number;
  augTotalHolidays: number;
  sepTotalHolidays: number;
  octTotalHolidays: number;
  novTotalHolidays: number;
  decTotalHolidays: number;


  janAvailabilityAfterHolidaysInDays: number;
  febAvailabilityAfterHolidaysInDays: number;
  marAvailabilityAfterHolidaysInDays: number;
  aprAvailabilityAfterHolidaysInDays: number;
  mayAvailabilityAfterHolidaysInDays: number;
  junAvailabilityAfterHolidaysInDays: number;
  julAvailabilityAfterHolidaysInDays: number;
  augAvailabilityAfterHolidaysInDays: number;
  sepAvailabilityAfterHolidaysInDays: number;
  octAvailabilityAfterHolidaysInDays: number;
  novAvailabilityAfterHolidaysInDays: number;
  decAvailabilityAfterHolidaysInDays: number;

}